clean up archive

+++ after_knit.sh
@@ -14,9 +14,13 @@ done
 cp -r docs/* archive
 rm -r archive/img archive/archive #archive/*/*.pdf
 tar -zcvf archive.tar.gz archive
-rm -r archive
 mv archive.tar.gz docs/
 #tar -xvzf archive.tar.gz
+
+# Clean up
+[[ -d archive ]] && rm -r archive/
+[[ -d docs/archive ]] && rm -r archive
+
 
 # Create index.md from list of directories
 touch docs/index.md
